Fixes #1427 (#1431)
The overview page component no longer needs to try
to fetch depending on the location. Since we run once and try to prevent
duplicate fetches, we did a fetch on the N-1 state.

Another bug I found while looking at this, if we refresh the page on a
global saved search, the details would not load up. This was due
to when the assignment of the first current global saved search.
It was actually returning undefined because it did not have the currentLocation.
This also fixes the problem where the shoelace toaster is appearing.

Other changes:
- Add test data to help test one of the Top XYZ bookmarks (But keep it at 80 features to prevent loading too much data)