Fix playwright test flakiness
During the saved search playwright tests, we kill and reopen the ports needed for the database (as well as all other ports).

We do this because beforehand, we did not port forward the ports for the database and only the ports needed to connect to the app (the backend already had access to the database internally)

But instead of doing that, we now port forward the database ports. Then we have a new helper makefile target to check if the port is live.

We use `nc` instead of curl because it did not work with grpc. The devcontainer Dockerfile has been updated to install that dependency (devs will need to rebuild their devcontainer)

Hopefully, this will reduce the times when the port fails to forward again on subsequent tries since we never have to forward it again.

One thing to note: Now developers will need to run make port-forward-manual first before dev_fake_data and other commands after running start-local