chore: add unstaged changes; detect unstaged changes (#1574)
Seems like the combination of multiple dependabot changes left package-lock.json with unstaged changes again.

This changes adds those.

Also, added a Makefile target to detect them called `unstaged-changes`. This target is used in the precommit target so we should catch this in the future as the dependabot PRs are merged into main.