Minimize valkey downtime by creating the new instance first
There was a recent change that causes valkey to recreate the instance.
By default, terraform will terminate the instance (10-15 mins) before
creating it (10-15 minutes).
When I deployed this week's release to staging, it was discovered that the app hangs
while still trying to connect to valkey and it messes up the response time
during the full 30 minutes of waiting.

This change minimizes valkey downtime by creating the
new instance before terminating the old one. To do this, we have to add
some random characters to the end of the name since there can only be
one instance with a given name per GCP project. Terraform will generate
a new random string the next time it has to create it. Now, there
should be minimal performance degradation since the new instance is active
and ready before the old one is terminated.

An added plus, this allows us to easily keep up to date with Valkey since
valkey 8.1 was released. Just waiting on GCP to support it.

In the future, we should also add a timeout to the context for the
caching layer. So that it will not hang and still return the result if
the caching layer fails. Previously, when there was a
single cache middleware layer, the app could handle this gracefully even
without the timeout (without us doing anything).
But when we moved to per endpoint caching, something happened.