feature: Add new spanner opeartors for search type
When it comes to searching by name and description, we actually
do a wildcard search. The operator for those two searches are EQ/NEQ
and we manually convert to LIKE and NOT LIKE inside searchOperatorToSpannerStringPatternOperator

This change adds two new operators: LIKE and NOT_LIKE and those are used
for searching by name and description during the ANTLR visitor code.