Change ANY_VALUE to use feature description identifier
Currently, when a user searches without a term, it defaults to using
the name identifier which does a wildcard search on the name.

This commit changes the behavior so that it uses the description identifier.

And since the spanner layer searches both the description and the name
using a wildcard search when it sees description identifier, I updated the search docs.

This change should happen on a separate deployment from this [PR](https://github.com/GoogleChrome/webstatus.dev/pull/1586)
because we need the description to be populated in the spanner database.
Otherwise, current users without the term will get zero results suddenly.