feat: Add FeatureGroupIDsLookup table for better feature search perf
Performance on the front page search begins to suffer when
there are multiple group terms separated by OR operators.

Upon inspection, it turns out that the existing `groupFilter`
is pretty complex because on the fly it tries to find the features not
only belonging to a group but the child groups too.

It would be better to calculate this data for better performance.

FeatureGroupIDsLookup contains that calculated data.