feat: add FeatureGroupKeysLookup table for better feature search perf
This change is a split up of #1630. Check that description for the high level reason.

This particular change is the first part where we add a new table that contains the data.

The spanner method `UpsertFeatureGroupLookups` contains the logic to gather all the data and populate the table.

Notes about the table:
- I added a `Depth` column to enable future use cases in case users want to know only the features at a given level and not all the features for children groups.