chore: Fix July 22 build (#1644)
Currently, the builds are broken because the go linter updated.

[Example Build](https://github.com/GoogleChrome/webstatus.dev/actions/runs/16449803280/job/46491701520)

In particular, the new lint rule is noctx. And it is saying to use slog.XYZContext instead of slog.XYZContext.
This change contains the fixes to those new linter problems.
I used the available ctx object for those if it was available. Otherwise, used the placeholder std lib context.TODO().

Other changes:
- Update the tools in devcontainer.json
- Run make go-update && make go-tidy to update the go dependencies
- Run `npm audit fix --force` to fix any npm vulnerabilities