Update docs about wrench, use version from tools/go.mod (#1694)
After #1674, wrench is accessed via `go tool wrench`. This change updates
the deployment documentation to reflect that.

Also, I saw that the wrench tool that the spanner emulator uses was outdated.

Now that we have the version in go.mod, I extract the version from the go.mod file
and use that in the spanner dockerfile. Now, we only have to maintain one
place for the version.