fix: Retry connection to valkey
Fixes #1698

NewClient tries to dial the valkey instance before returning.
If there's a problem with dialing valkey, it will fail and not recover.
As a result, the cloud run instance will fail to deploy.

This change adds retry logic in an attempt to give the server more attempts
before failing.