feat: Workflow handles for Moved and Split features
This commit builds on the previous changes by having the workflow call the glue layer
to store moved and split features.

The `WebFeatureStorer` interface is extended with `InsertMovedWebFeatures` and
`InsertSplitWebFeatures` methods. The main job processor in the workflow now calls these new methods.

In the future, we will intrdouce the spanner tables and implement the methods in the spanner layer that will actually be used.

Future TODO: DRY up the code in web_features_worker_test.go