feat: Add placeholder spanner methods for new tables
This change adds placeholder methods for the spanner layer.

There are two sync methods. These will be used during the ingestion of the web features data.
Then there are two `Get` methods which will be used during `/v1/features/{featureid}` in the event the feature key does not exist in the existing Web Features Table.