feat: Handle moved features during data ingestion (#1769)
This change introduces logic to handle features that have been marked as "moved" to a new feature key.

During data ingestion for WPT and Chromium histograms, if a feature has been moved, this change ensures that the metrics are associated with the new, redirected feature key. This allows for future ingestions to be correctly attributed while upstream sources (like WPT test annotations or Chromium mojom files) are being updated.

These checks are implemented at different layers depending on the data source:

- For WPT, the migration happens in the workflow layer before scoring. This allows the scores to be calculated with the corrected feature keys.
- For Chromium histograms, the migration happens in the spanner adapter layer. This is because the full list of feature keys is not available until this layer, where it is fetched from the database.

The implementation also introduces a check to prevent conflicts where both the old and new feature keys are present in the source data. When this conflict occurs, an error is logged with the message "conflict migrating feature key", which will be used for alerting in GCP.

We can also add alerting for these messages too:
- "migrating feature key for histogram"
- "migrating feature key for test"

These will let us know to update the upstream sources