refactor: Remove repo-downloader service and GCS emulator (#1806)
This commit removes the repo-downloader service and the associated local GCS emulator. The service was responsible for downloading git repositories, and this functionality is no longer required.

The repo downloader was first used when repos were not publishing artifacts. And instead, we would need to download them ourselves. However, we have not had to do that in over a year now. We can safely
delete this code now.

Key changes:

- Deleted the Go source code, Terraform configuration, and OpenAPI specification for the repo-downloader service.
- Removed the local GCS emulator from the development environment.
- Updated the Makefile, development documentation, and VSCode launch configuration to remove all references to the deleted components.