refactor: Use sync for browser feature availabilities
This commit refactors the code to use a new sync method for updating
browser feature availabilities. The new method, SyncBrowserFeatureAvailabilities,
allows for batch updates, which improves performance by reducing the number
of database calls. And also helps in the cases when a feature's availability regresses for a browser.

The following changes were made:
- Replaced UpsertBrowserFeatureAvailability with SyncBrowserFeatureAvailabilities
  in all the test files.
- Updated the main logic to use the new sync method.
- Updated the data loading scripts to use the new sync method.