Automatically create deployment issue
When we deploy to production we run the script create-deplyoment-issue
to create the github issue for the deployment. This makes that step
automatic, when running the deploy-production script we'll automatically
create that issue if it doesn't already exist, and if it exists and has
already been closed, we'll abort the deployment.