chore: update spanner package (#1942)
We were using an old version of the spanner package. However, we need the newer version due to
https://github.com/googleapis/google-cloud-go/commit/003abca9172082ad1f2fbcc9b37639f389ade8ee#diff-ffa3ef304f4803a0bb8d3cc088fe526420f8f87a4f205b67e7bb29936bb5307d

And there are some upcoming changes that could benefit from that.

Removed the lines to exclude newer versions of the SDK and add env var flags to disable multiplex sessions.

Filed https://github.com/GoogleChrome/webstatus.dev/issues/1943 to re-enable multiplexed sessions once the underlying issue is fixed. More details in that bug.