feat(infra): Shorten workflow names to accommodate service account limits
Shortens the 'short_name' of the web-features-mapping workflow to 'feature-map'.

Also adds validation to the 'single_stage_go_workflow' module to enforce a maximum length of 13 characters for the 'short_name'.

This is necessary because the service account name is generated from the short_name and includes the environment, and service account names have a limited length.